Refactored remaining operators into specific modules Conflicts: 	io_scene_nif/operators.py
