+++ io_scene_nif/operators/__init__.py
@@ -37,5 +37,5 @@
 #
 # ***** END LICENSE BLOCK *****
 
-from . import object
+from . import object, geometry, nif_import_op, nif_export_op, nif_common_op
