+++ io_scene_nif/operators/object.py
